requirements.txt
flask
scikit-learn
joblib
pillow
gunicorn
